EconomicChoice

Documentation for EconomicChoice.

EconomicChoice.choice_summary

EconomicChoice.regress_behavior

EconomicChoice.regress_variable

EconomicChoice.choice_summary — Method

choice_summary(trials)

Return a dictionary mapping each offer type to a MeanStd object containing the mean and standard deviation of the fraction of A choices for that offer.

EconomicChoice.regress_behavior — Method

regress_behavior(trials)
regress_behavior(offerscalar, trials)

Fit a logistic regression model to the behavior of a subject. trials is a vector of TwoJuiceTrial objects.

Optionally pass offerscalar to specify a function that computes the offer scalar (the independent variable) from the number of juice drops; the default is to use log(nB/nA).

EconomicChoice.regress_variable — Method

regress_variable(f, trials, responses)

Fit a linear regression model to neural responses. f is a function that computes the independent variable from the trials. trials is a vector of TwoJuiceTrial objects.